requirements.txt
gradio
matplotlib
numpy
scikit-learn
openai
python-dotenv